+++ Neironikus/neuron.h
@@ -2,7 +2,7 @@
 class neuron //Класс нейрон 
 {
 public:
-	float weight;
-	float input[2];
-	float output(float in[]);
+	double weight;
+	double input[2];
+	double output(double in[]);
 };
